requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
decorator==5.1.1
Flask==2.2.2
Flask-CORS==4.0.0
flatlib==0.2.3
future==0.18.3
geocoder==1.38.1
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.0
pyparsing==3.0.9
pyswisseph==2.8.0.post1
python-dateutil==2.8.2
ratelim==0.1.6
requests==2.28.2
six==1.16.0
urllib3==1.26.14
Werkzeug==2.2.2
